+++ nuxt.config.js
@@ -12,8 +12,8 @@ export default {
       { src: '/js/jquery.min.js' }
     ],
     link: [
-      { 
-        rel: 'stylesheet', 
+      {
+        rel: 'stylesheet',
         href: 'https://use.fontawesome.com/releases/v5.7.2/css/all.css',
         integrity: "sha384-fnmOCqbTlWIlj8LyTjo7mOUStjsKC4pOpQbqyi7RrhN7udi9RwhKkMHpvLbHG9Sr",
         crossorigin: "anonymous",
@@ -33,11 +33,13 @@ export default {
     '@nuxtjs/proxy',
     '@nuxtjs/toast',
     ['nuxt-vuex-localstorage', {
-      localStorage: ['user']}]
+      localStorage: ['user']}
+    ],
+    'nuxt-client-init-module',
   ],
   proxy: {
     '/graphql': {
-      target: 'https://api.qetela.tk',
+      target: 'http://127.0.0.1:8000',
       pathRewrite: {
         '^/graphql' : '/graphql'
         }
@@ -60,7 +62,7 @@ export default {
     clientConfigs: {
       default: {
         // required
-        httpEndpoint: 'https://api.qetela.tk/graphql',
+        httpEndpoint: 'http://127.0.0.1:8000/graphql',
         // optional
         // See https://www.apollographql.com/docs/link/links/http.html#options
         httpLinkOptions: {
@@ -78,7 +80,7 @@ export default {
         websocketsOnly: false // Optional
       },
       test: {
-        httpEndpoint: 'https://api.qetela.tk/graphql',
+        httpEndpoint: 'http://127.0.0.1:8000/graphql',
         wsEndpoint: null,//'ws://127.0.0.1:8000/graphql',
         tokenName: 'apollo-token'
       },